Integratorでパース処理のテストを有効化

+++ 10/integrator_test.go
@@ -63,6 +63,34 @@ func TestIntegratorIntegrate(t *testing.T) {
 				"Fixture/ArrayTest/cmp/Main.xml",
 			},
 		},
+		{
+			desc: "Square",
+			src: []string{
+				"Fixture/Square/Main.jack",
+				"Fixture/Square/Square.jack",
+				"Fixture/Square/SquareGame.jack",
+			},
+			destTokenizedXML: []string{
+				"Fixture/Square/MainT.xml",
+				"Fixture/Square/SquareT.xml",
+				"Fixture/Square/SquareGameT.xml",
+			},
+			destParsedXML: []string{
+				"Fixture/Square/Main.xml",
+				"Fixture/Square/Square.xml",
+				"Fixture/Square/SquareGame.xml",
+			},
+			wantTokenizedXML: []string{
+				"Fixture/Square/cmp/MainT.xml",
+				"Fixture/Square/cmp/SquareT.xml",
+				"Fixture/Square/cmp/SquareGameT.xml",
+			},
+			wantParsedXML: []string{
+				"Fixture/Square/cmp/Main.xml",
+				"Fixture/Square/cmp/Square.xml",
+				"Fixture/Square/cmp/SquareGame.xml",
+			},
+		},
 	}
 
 	for _, tc := range cases {
@@ -88,16 +116,15 @@ func TestIntegratorIntegrate(t *testing.T) {
 				}
 			}
 
-			for _, dest := range tc.destParsedXML {
-				//got := readFileQuietly(dest)
-				//want := readFileQuietly(tc.wantParsedXML[i])
-				//
-				//if diff := cmp.Diff(got, want); diff != "" {
-				//	t.Errorf("failed: diff %s %s: (-got +want):\n%s\n", dest, tc.wantTokenizedXML[i], diff)
-				//} else {
-				//	os.Remove(dest)
-				//}
-				os.Remove(dest)
+			for i, dest := range tc.destParsedXML {
+				got := readFileQuietly(dest)
+				want := readFileQuietly(tc.wantParsedXML[i])
+
+				if diff := cmp.Diff(got, want); diff != "" {
+					t.Errorf("failed: diff %s %s: (-got +want):\n%s\n", dest, tc.wantTokenizedXML[i], diff)
+				} else {
+					os.Remove(dest)
+				}
 			}
 		})
 	}
@@ -139,45 +166,6 @@ func TestIntegratorIntegrateFile(t *testing.T) {
 				os.Remove(tc.destTokenizedXML)
 			}
 
-			//gotParsedXML := readFileQuietly(tc.destParsedXML)
-			//wantParsedXML := readFileQuietly(tc.wantParsedXML)
-			//
-			//if diff := cmp.Diff(gotParsedXML, wantParsedXML); diff != "" {
-			//	t.Errorf("failed: diff %s %s: (-got +want):\n%s\n", tc.destParsedXML, tc.wantParsedXML, diff)
-			//} else {
-			//	os.Remove(tc.destParsedXML)
-			//}
-			os.Remove(tc.destParsedXML)
-		})
-	}
-}
-
-// TODO パース処理が一通り実装できるまで、途中状態のXMLファイルをリグレッションテストできるようにしておく
-func TestIntegratorProvisional(t *testing.T) {
-	cases := []struct {
-		desc             string
-		src              string
-		destTokenizedXML string
-		destParsedXML    string
-		wantParsedXML    string
-	}{
-		{
-			desc:             "途中状態のXMLファイルをリグレッションテスト",
-			src:              "Fixture/ExpressionLessSquare/Main.jack",
-			destTokenizedXML: "Fixture/ExpressionLessSquare/MainT.xml",
-			destParsedXML:    "Fixture/ExpressionLessSquare/Main.xml",
-			wantParsedXML:    "Fixture/ExpressionLessSquare/provisional/Main.xml",
-		},
-	}
-
-	for _, tc := range cases {
-		os.Remove(tc.destTokenizedXML)
-		os.Remove(tc.destParsedXML)
-
-		t.Run(tc.desc, func(t *testing.T) {
-			integrator := NewIntegrator([]string{})
-			integrator.integrateFile(tc.src)
-
 			gotParsedXML := readFileQuietly(tc.destParsedXML)
 			wantParsedXML := readFileQuietly(tc.wantParsedXML)
 
@@ -186,7 +174,6 @@ func TestIntegratorProvisional(t *testing.T) {
 			} else {
 				os.Remove(tc.destParsedXML)
 			}
-			os.Remove(tc.destTokenizedXML)
 		})
 	}
 }
